+++ app/main/forms.py
@@ -1,13 +1,13 @@
 # -*- coding:utf-8 -*-
 
 from flask_wtf import Form
-# from flask_wtf.file import FileField, FileAllowed, FileRequired
+from flask_wtf.file import FileField, FileAllowed, FileRequired
 from wtforms import StringField, TextAreaField, SelectField, SubmitField
 from wtforms import ValidationError
 from wtforms.validators import Required, Length, Email, Regexp
 from flask_pagedown.fields import PageDownField
 from ..models import Role, User
-# from .. import Images
+from .. import photos
 
 class NameForm(Form):
     name = StringField('name', validators=[Required()])
@@ -60,8 +60,8 @@ class CommentForm(Form):
     submit = SubmitField('Submit')
 
 
-# class PhotoForm(Form):
-    # photo = FileField('Upload Image', validators=[
-        # FileRequired(),
-        # FileAllowed(Images.images, 'Image Only')])
-    # submit = SubmitField('Submit')
+class PhotoForm(Form):
+    photo = FileField('Upload Image', validators=[
+        FileRequired(),
+        FileAllowed(photos, 'Image Only')])
+    submit = SubmitField('Submit')
