+++ app/templates/base.html
@@ -35,7 +35,11 @@
                 {% if current_user.is_authenticated %}
                 <li class="dropdown">
                   <a href="#" class="dropdown-toggle" data-toggle="dropdown">
+                      {% if current_user.avatar_url %}
+                      <img src="{{ current_user.avatar_url }}" style="width:18px;height:18px">
+                      {% else %}
                       <img src="{{ current_user.gravatar(size=18) }}">
+                      {% endif %}
                       Account <b class="caret"></b></a>
                   <ul class="dropdown-menu">
                     <li><a href="{{ url_for('auth.change_password') }}">Change Password</a></li>
