+++ app/templates/user.html
@@ -5,7 +5,11 @@
 
 {% block page_content %}
 <div class="page-header">
+    {% if not user.avatar_url %}
     <img class="img-rounded profile-thumbnail" src="{{ user.gravatar(size=256) }}">
+    {% else %}
+    <img class="img-rounded profile-thumbnail" src="{{ user.avatar_url }}" style="width:256px;height:256px;">
+    {% endif %}
     <div class="profile-header">
     <h1>{{ user.username }}</h1>
     {% if user.name or user.location %}
@@ -31,6 +35,9 @@
     {% if user == current_user %}
     <a class="btn btn-default" href="{{ url_for('.edit_profile') }}">
         Edit Profile
+    </a>
+    <a class="btn btn-default" href="{{ url_for('.setting_avatar') }}">
+        Upload Photo
     </a>
     {% endif %}
     {% if current_user.is_administrator() %}
