+++ config.py
@@ -6,11 +6,13 @@ class Config:
     SSL_DISABLE = False
     SQLALCHEMY_TRACK_MODIFICATIONS = True
     SQLALCHEMY_COMMIT_ON_TEARDOWN = True
+    # mail
     MAIL_SERVER = 'smtp.163.com'
     MAIL_PORT = 25
     MAIL_USE_TLS = True
     MAIL_USERNAME = os.environ.get('MAIL_USERNAME')
     MAIL_PASSWORD = os.environ.get('MAIL_PASSWORD')
+    # send mail
     FLASKY_MAIL_SUBJECT_PREFIX = '[Hello]'
     FLASKY_MAIL_SENDER = 'Flasky Admin <18292878089@163.com>'
     FLASKY_ADMIN = os.environ.get('FLASKY_ADMIN')
@@ -19,6 +21,8 @@ class Config:
     FLASKY_COMMENTS_PER_PAGE = 30
     FLASKY_SLOW_DB_QUERY_TIME=0.5
     FLASKY_COMMENTS_PER_PAGE = 10
+
+    UPLOADS_DEFAULT_DEST = os.path.join(basedir, 'static/upload')
 
     @staticmethod
     def init_app(app):
